requirements.txt
streamlit
onnx
onnxruntime
numpy
opencv-python-headless
Pillow
